Implement in java the following methods:
- Display data for all classes.
- Display data for a course with specified id.
- Display data for all classes whose title (when converted to all
lowercase letters) begins with "intro".
- Display data for all classes whose dept (when converted to all
lowercase letters) is "cos" and whose coursenum begins with "3".

considering the following SQL database:

The school database consists of these five tables:

1. classes
Fields: classid, courseid, days, starttime, endtime, bldg, roomnum

2. courses
Fields: courseid, area, title, descrip, prereqs
The courseid field of the courses table is related to the courseid field
of the classes table.

3. crosslistings
Fields: courseid, dept, coursenum
The courseid field of the crosslistings table is related to the courseid
field of the courses table.

4. coursesprofs
Fields: courseid, profid
The courseid field of the coursesprofs table is related to the courseid
field of the courses table.

5. profs
Fields: profid, profname
The profid field of the profs table is related to the profid field of
the coursesprofs table.